Fixed Player Movement + Recreated the test Scene
When a player stands on a Plattform he now gets parented to a new Gameobject. This Gameobject is a child of the parent of the Platform.
The Player can now move freely on plattforms cause the Plattforms executed their movement in an Update Function and not in an FixedUpdate Function. I also set the Fixed Timestamp in ProjectSettings -> Time to 0.005.
And a Gizmos was added to the Rotationg Plattform