Player Movement and Cam movement Seperate + Player Prefab
The Player Movement and the Camera Movement are sperated now and it plays like you would expect from a normal third person game. We now have a player Prefab, although if you use it dont forget to attach a CinemachineBrain Component to the Main Camera. If you dont do that, im pretty sure the camera wont work. Ah i forgot to mention that i added the cinenachine Package from unity with which you can control the camera in various ways and i also needed it for the new camera behaviour. I still plan on adding scrolling so that you can zoom in/out and maybe a acceration/decelartion of the player movement.